requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4

# AI และ Computer Vision
torch==2.3.0
numpy==1.24.4
onnx==1.15.0
onnxruntime-gpu==1.15.1
insightface==0.7.1
Pillow==10.1.0
opencv-python==4.8.1.78
scikit-image==0.21.0
scipy==1.11.3

# Parallel Processing
ray==2.9.0

# Database
pymilvus==2.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1